requirements.txt
streamlit>=1.28.0
openai>=1.3.0
transformers>=4.30.0
sentence-transformers>=2.2.0
scikit-learn>=1.2.0
pandas>=1.5.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.15.0
nltk>=3.8.0
gensim>=4.2.0
torch>=1.13.0
wordcloud>=1.9.0
umap-learn>=0.5.0
python-dotenv>=1.0.0
joblib>=1.2.0
datasets==2.14.0 